backend/pyproject.toml
[project]
name = "scoop-easy-backend"
version = "0.1.0"
description = "Backend API for Scoop Easy - A GUI tool for Scoop package manager"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "pywebview>=5.0",
]
